fix: migrations error fixing uuid into ID

+++ database/migrations/1745617085963_journals.ts
@@ -2,17 +2,26 @@ import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class Journals extends BaseSchema {
   protected tableName = 'journals'
+
   public async up() {
-    this.schema.createTable(this.tableName, (table) => {
-      table.increments('id')
-      table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
-      table.string('title').notNullable()
-      table.text('content').notNullable()
-      table.date('entry_date').notNullable()
-      table.timestamps(true)
-    })
+    const exists = await this.schema.hasTable(this.tableName)
+    if (!exists) {
+      this.schema.createTable(this.tableName, (table) => {
+        table.increments('id').primary()
+        table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
+        table.string('title').notNullable()
+        table.text('content').notNullable()
+        table.date('entry_date').notNullable()
+        table.timestamp('created_at', { useTz: true }).defaultTo(this.now())
+        table.timestamp('updated_at', { useTz: true }).defaultTo(this.now())
+      })
+    }
   }
+
   public async down() {
-    this.schema.dropTable(this.tableName)
+    const exists = await this.schema.hasTable(this.tableName)
+    if (exists) {
+      this.schema.dropTable(this.tableName)
+    }
   }
 }
